All the programming to calculate and analyze the data its "ready"

+++ trunk/src/GUIFramework/CommonConfigWindows/DataAnalyzeWindow/CurveData.h
@@ -158,7 +158,7 @@ private:
 	double			mMaxHeight;
 	double			mMinHeight;
 
-	int				mFp;
+	double			mFp;
 	double			mHs;
 	double 			mH;
 	double			mTp;
